Update codebase to upstream revision 545d40e3141ea73ac871b81d943e4b2cad5fe3b2.

+++ tests/subsys/llext/src/test_llext.c
@@ -523,7 +523,8 @@ ZTEST(llext, test_find_section)
 	res = llext_get_section_header(loader, ext, ".data", &shdr);
 	zassert_ok(res, "get_section_header() should succeed");
 	zassert_equal(shdr.sh_offset, section_ofs,
-		     "different section offset %zd from get_section_header", shdr.sh_offset);
+		     "different section offset %llu from get_section_header",
+		     (uint64_t)shdr.sh_offset);
 
 	uintptr_t symbol_ptr = (uintptr_t)llext_find_sym(&ext->exp_tab, "number");
 	uintptr_t section_ptr = (uintptr_t)find_section_ext + section_ofs;
